refactor: Move promptForAppMap to its own file
