+++ src/lib/AppMapQuickPickItem.ts
@@ -1,0 +1,5 @@
+import * as vscode from 'vscode';
+
+export type AppMapQuickPickItem = vscode.QuickPickItem & {
+  resourceUri: vscode.Uri;
+};
